AUDIO UPDATE
not sure why it was till in my branch but it seems the changes i did last time where not sent to the commit and pull request. When i cinnected to my mac to check all the work i have done wasn't there so now i understand why. I might have done a mistake.